Refactor the webui integration test to support external image
This change refactors the integration test so that it can be
run using a local image or an external image. It adds a test/e2e.sh
script for running the test which handles the details of pushing
images to integrated/external registries or allowing the image
to be referenced directly from the host in the case of 'oc cluster up'.